Separating comments functions from file process stage to comments stage
