+++ src/core/infrastructure/adapters/services/codeBase/codeReviewPipeline/stages/create-file-comments.stage.ts
@@ -1,0 +1,536 @@
+import { Inject, Injectable } from '@nestjs/common';
+import { BasePipelineStage } from '../../../pipeline/base-stage.abstract';
+import { CodeReviewPipelineContext } from '../context/code-review-pipeline.context';
+import { PinoLoggerService } from '../../../logger/pino.service';
+import {
+    ClusteringType,
+    CodeReviewConfig,
+    CodeSuggestion,
+    CommentResult,
+    FileChange,
+    Repository,
+} from '@/config/types/general/codeReview.type';
+import { OrganizationAndTeamData } from '@/config/types/general/organizationAndTeamData';
+import {
+    COMMENT_MANAGER_SERVICE_TOKEN,
+    ICommentManagerService,
+} from '@/core/domain/codeBase/contracts/CommentManagerService.contract';
+import {
+    IPullRequestsService,
+    PULL_REQUESTS_SERVICE_TOKEN,
+} from '@/core/domain/pullRequests/contracts/pullRequests.service.contracts';
+import {
+    ISuggestionService,
+    SUGGESTION_SERVICE_TOKEN,
+} from '@/core/domain/codeBase/contracts/SuggestionService.contract';
+import { CodeManagementService } from '../../../platformIntegration/codeManagement.service';
+import { DeliveryStatus } from '@/core/domain/pullRequests/enums/deliveryStatus.enum';
+import { ImplementationStatus } from '@/core/domain/pullRequests/enums/implementationStatus.enum';
+import { AutomationStatus } from '@/core/domain/automation/enums/automation-status';
+import { AutomationExecutionEntity } from '@/core/domain/automation/entities/automation-execution.entity';
+import {
+    AUTOMATION_EXECUTION_SERVICE_TOKEN,
+    IAutomationExecutionService,
+} from '@/core/domain/automation/contracts/automation-execution.service';
+import { PlatformType } from '@/shared/domain/enums/platform-type.enum';
+import { PullRequestsEntity } from '@/core/domain/pullRequests/entities/pullRequests.entity';
+import { PullRequestReviewComment } from '@/core/domain/platformIntegrations/types/codeManagement/pullRequests.type';
+
+@Injectable()
+export class CreateFileCommentsStage extends BasePipelineStage<CodeReviewPipelineContext> {
+    readonly stageName = 'CreateFileCommentsStage';
+
+    constructor(
+        private readonly logger: PinoLoggerService,
+
+        @Inject(AUTOMATION_EXECUTION_SERVICE_TOKEN)
+        private readonly automationExecutionService: IAutomationExecutionService,
+
+        @Inject(COMMENT_MANAGER_SERVICE_TOKEN)
+        private readonly commentManagerService: ICommentManagerService,
+
+        @Inject(PULL_REQUESTS_SERVICE_TOKEN)
+        private readonly pullRequestService: IPullRequestsService,
+
+        @Inject(SUGGESTION_SERVICE_TOKEN)
+        private readonly suggestionService: ISuggestionService,
+
+        private readonly codeManagementService: CodeManagementService,
+    ) {
+        super();
+    }
+
+    protected async executeStage(
+        context: CodeReviewPipelineContext,
+    ): Promise<CodeReviewPipelineContext> {
+        // Validações fundamentais de segurança
+        if (!context?.organizationAndTeamData) {
+            this.logger.error({
+                message: 'Missing organizationAndTeamData in context',
+                context: this.stageName,
+            });
+            return context;
+        }
+
+        if (!context?.pullRequest?.number) {
+            this.logger.error({
+                message: 'Missing pullRequest data in context',
+                context: this.stageName,
+                metadata: { organizationAndTeamData: context.organizationAndTeamData },
+            });
+            return context;
+        }
+
+        if (!context?.repository?.name || !context?.repository?.id) {
+            this.logger.error({
+                message: 'Missing repository data in context',
+                context: this.stageName,
+                metadata: {
+                    organizationAndTeamData: context.organizationAndTeamData,
+                    prNumber: context.pullRequest.number,
+                },
+            });
+            return context;
+        }
+
+        // Verificar se há sugestões para processar
+        const validSuggestions = context?.validSuggestions || [];
+        const discardedSuggestions = context?.discardedSuggestions || [];
+        const overallComments = context?.overallComments || [];
+        const changedFiles = context?.changedFiles || [];
+
+        if (validSuggestions.length === 0) {
+            this.logger.log({
+                message: `No file-level suggestions to process for PR#${context.pullRequest.number}`,
+                context: this.stageName,
+                metadata: {
+                    organizationAndTeamData: context.organizationAndTeamData,
+                    prNumber: context.pullRequest.number,
+                    discardedSuggestionsCount: discardedSuggestions.length,
+                },
+            });
+            return this.updateContext(context, (draft) => {
+                draft.lineComments = [];
+                draft.lastAnalyzedCommit = null;
+            });
+        }
+
+        try {
+            this.logger.log({
+                message: `Starting file comments creation for PR#${context.pullRequest.number}`,
+                context: this.stageName,
+                metadata: {
+                    validSuggestionsCount: validSuggestions.length,
+                    discardedSuggestionsCount: discardedSuggestions.length,
+                    overallCommentsCount: overallComments.length,
+                    organizationAndTeamData: context.organizationAndTeamData,
+                    prNumber: context.pullRequest.number,
+                },
+            });
+
+            const { lineComments, lastAnalyzedCommit } = await this.finalizeReviewProcessing(
+                context,
+                changedFiles,
+                validSuggestions,
+                discardedSuggestions,
+                overallComments,
+            );
+
+            this.logger.log({
+                message: `Successfully processed file comments for PR#${context.pullRequest.number}`,
+                context: this.stageName,
+                metadata: {
+                    lineCommentsCreated: lineComments.length,
+                    organizationAndTeamData: context.organizationAndTeamData,
+                    prNumber: context.pullRequest.number,
+                },
+            });
+
+            return this.updateContext(context, (draft) => {
+                draft.lineComments = lineComments;
+                draft.lastAnalyzedCommit = lastAnalyzedCommit;
+            });
+
+        } catch (error) {
+            this.logger.error({
+                message: `Error during file comments creation for PR#${context.pullRequest.number}`,
+                context: this.stageName,
+                error,
+                metadata: {
+                    organizationAndTeamData: context.organizationAndTeamData,
+                    validSuggestionsCount: validSuggestions.length,
+                },
+            });
+
+            // Em caso de erro, retorna contexto com valores padrão
+            return this.updateContext(context, (draft) => {
+                draft.lineComments = [];
+                draft.lastAnalyzedCommit = null;
+            });
+        }
+    }
+
+    /**
+     * Finalizes the code review process by generating comments and saving suggestions
+     * @param context Pipeline context
+     * @param changedFiles Files changed in the PR
+     * @param validSuggestionsToAnalyze Valid suggestions found
+     * @param discardedSuggestionsBySafeGuard Discarded suggestions
+     * @param overallComments General comments
+     * @returns Processing result with comments and suggestions
+     */
+    private async finalizeReviewProcessing(
+        context: CodeReviewPipelineContext,
+        changedFiles: FileChange[],
+        validSuggestionsToAnalyze: Partial<CodeSuggestion>[],
+        discardedSuggestionsBySafeGuard: Partial<CodeSuggestion>[],
+        overallComments: { filepath: string; summary: string }[],
+    ): Promise<{
+        lineComments: Array<CommentResult>;
+        lastAnalyzedCommit: any;
+    }> {
+        const {
+            organizationAndTeamData,
+            pullRequest,
+            codeReviewConfig,
+            repository,
+            platformType,
+        } = context;
+
+        // Sort and prioritize suggestions
+        const { sortedPrioritizedSuggestions, allDiscardedSuggestions } =
+            await this.suggestionService.sortAndPrioritizeSuggestions(
+                organizationAndTeamData,
+                codeReviewConfig,
+                pullRequest,
+                validSuggestionsToAnalyze,
+                discardedSuggestionsBySafeGuard,
+            );
+
+        // Create line comments
+        const { commentResults, lastAnalyzedCommit } =
+            await this.createLineComments(
+                organizationAndTeamData,
+                pullRequest,
+                sortedPrioritizedSuggestions,
+                repository,
+                codeReviewConfig,
+                platformType,
+            );
+
+        // Save pull request suggestions
+        await this.savePullRequestSuggestions(
+            organizationAndTeamData,
+            pullRequest,
+            repository,
+            changedFiles,
+            commentResults,
+            sortedPrioritizedSuggestions,
+            allDiscardedSuggestions,
+            platformType,
+            context.fileMetadata,
+        );
+
+        // Resolve comments that refer to suggestions partially or fully implemented
+        await this.resolveCommentsWithImplementedSuggestions({
+            organizationAndTeamData,
+            repository,
+            prNumber: pullRequest.number,
+            platformType: platformType as PlatformType,
+        });
+
+        return {
+            lineComments: commentResults,
+            lastAnalyzedCommit,
+        };
+    }
+
+    private calculateStartLine(suggestion: any) {
+        if (
+            suggestion.relevantLinesStart === undefined ||
+            suggestion.relevantLinesStart === suggestion.relevantLinesEnd
+        ) {
+            return undefined;
+        }
+        return suggestion.relevantLinesStart + 15 > suggestion.relevantLinesEnd
+            ? suggestion.relevantLinesStart
+            : undefined;
+    }
+
+    private calculateEndLine(suggestion: any) {
+        if (
+            suggestion.relevantLinesStart === undefined ||
+            suggestion.relevantLinesStart === suggestion.relevantLinesEnd
+        ) {
+            return suggestion.relevantLinesEnd;
+        }
+        return suggestion.relevantLinesStart + 15 > suggestion.relevantLinesEnd
+            ? suggestion.relevantLinesEnd
+            : suggestion.relevantLinesStart;
+    }
+
+    private async createLineComments(
+        organizationAndTeamData: OrganizationAndTeamData,
+        pullRequest: { number: number },
+        sortedPrioritizedSuggestions: any[],
+        repository: Partial<Repository>,
+        codeReviewConfig: CodeReviewConfig,
+        platformType: string,
+    ) {
+        try {
+            const lineComments = sortedPrioritizedSuggestions
+                .filter(
+                    (suggestion) =>
+                        suggestion.clusteringInformation?.type !==
+                        ClusteringType.RELATED,
+                )
+                .map((suggestion) => ({
+                    path: suggestion.relevantFile,
+                    body: {
+                        language: repository?.language,
+                        improvedCode: suggestion?.improvedCode,
+                        suggestionContent: suggestion?.suggestionContent,
+                        actionStatement:
+                            suggestion?.clusteringInformation
+                                ?.actionStatement || '',
+                    },
+                    start_line: this.calculateStartLine(suggestion),
+                    line: this.calculateEndLine(suggestion),
+                    side: 'RIGHT',
+                    suggestion,
+                }));
+
+            const { lastAnalyzedCommit, commentResults } =
+                await this.commentManagerService.createLineComments(
+                    organizationAndTeamData,
+                    pullRequest?.number,
+                    {
+                        name: repository.name,
+                        id: repository.id,
+                        language: repository.language,
+                    },
+                    lineComments,
+                    codeReviewConfig?.languageResultPrompt,
+                );
+
+            return { lastAnalyzedCommit, commentResults };
+        } catch (error) {
+            this.logger.log({
+                message: `Error when trying to create line comments for PR#${pullRequest.number}`,
+                error: error,
+                context: this.stageName,
+                metadata: {
+                    organizationAndTeamData,
+                    pullRequest,
+                    sortedPrioritizedSuggestions,
+                    repository,
+                },
+            });
+
+            const lastExecution =
+                await this.findLastTeamAutomationCodeReviewExecution(
+                    organizationAndTeamData.teamId,
+                    pullRequest.number,
+                    repository.id,
+                    platformType,
+                );
+
+            return {
+                lastAnalyzedCommit:
+                    lastExecution?.dataExecution?.lastAnalyzedCommit,
+                commentResults: [],
+            };
+        }
+    }
+
+    private async findLastTeamAutomationCodeReviewExecution(
+        teamAutomationId: string,
+        pullRequestNumber: number,
+        repositoryId: string,
+        platformType: string,
+    ) {
+        const lastTeamAutomationCodeReviewExecution: AutomationExecutionEntity =
+            await this.automationExecutionService.findLatestExecutionByDataExecutionFilter(
+                { pullRequestNumber: pullRequestNumber, platformType },
+                {
+                    status: AutomationStatus.SUCCESS,
+                    teamAutomation: { uuid: teamAutomationId },
+                    pullRequestNumber: pullRequestNumber,
+                    repositoryId: repositoryId,
+                },
+            );
+
+        return lastTeamAutomationCodeReviewExecution;
+    }
+
+    private async savePullRequestSuggestions(
+        organizationAndTeamData: OrganizationAndTeamData,
+        pullRequest: { number: number },
+        repository: Partial<Repository>,
+        changedFiles: FileChange[],
+        commentResults: CommentResult[],
+        sortedPrioritizedSuggestions: Partial<CodeSuggestion>[],
+        discardedSuggestions: Partial<CodeSuggestion>[],
+        platformType: string,
+        fileMetadata?: Map<string, any>,
+    ) {
+        const enrichedFiles = changedFiles.map((file) => {
+            const metadata = fileMetadata?.get(file.filename);
+            if (metadata) {
+                return {
+                    ...file,
+                    reviewMode: metadata.reviewMode,
+                    codeReviewModelUsed: metadata.codeReviewModelUsed,
+                };
+            }
+            return file;
+        });
+
+        const suggestionsWithStatus =
+            await this.suggestionService.verifyIfSuggestionsWereSent(
+                organizationAndTeamData,
+                pullRequest,
+                sortedPrioritizedSuggestions,
+                commentResults,
+            );
+
+        const pullRequestCommits =
+            await this.codeManagementService.getCommitsForPullRequestForCodeReview(
+                {
+                    organizationAndTeamData,
+                    repository: { id: repository.id, name: repository.name },
+                    prNumber: pullRequest.number,
+                },
+            );
+
+        await this.pullRequestService.aggregateAndSaveDataStructure(
+            pullRequest,
+            repository,
+            enrichedFiles,
+            suggestionsWithStatus,
+            discardedSuggestions,
+            platformType,
+            organizationAndTeamData?.organizationId,
+            pullRequestCommits,
+        );
+    }
+
+    private async resolveCommentsWithImplementedSuggestions({
+        organizationAndTeamData,
+        repository,
+        prNumber,
+        platformType,
+    }: {
+        organizationAndTeamData: OrganizationAndTeamData;
+        repository: Partial<Repository>;
+        prNumber: number;
+        platformType: PlatformType;
+    }) {
+        const codeManagementRequestData = {
+            organizationAndTeamData,
+            repository: {
+                id: repository.id,
+                name: repository.name,
+            },
+            prNumber: prNumber,
+        };
+
+        let isPlatformTypeGithub: boolean =
+            platformType === PlatformType.GITHUB;
+
+        const pr = await this.pullRequestService.findByNumberAndRepository(
+            prNumber,
+            repository.name,
+            organizationAndTeamData,
+        );
+
+        let implementedSuggestionsCommentIds =
+            this.getImplementedSuggestionsCommentIds(pr);
+
+        let reviewComments = [];
+
+        /**
+         * Marking comments as resolved in github needs to be done using another API.
+         * Marking comments as resolved in github also is done using threadId rather than the comment Id.
+         */
+        if (isPlatformTypeGithub) {
+            reviewComments =
+                await this.codeManagementService.getPullRequestReviewThreads(
+                    codeManagementRequestData,
+                    PlatformType.GITHUB,
+                );
+        } else {
+            reviewComments =
+                await this.codeManagementService.getPullRequestReviewComments(
+                    codeManagementRequestData,
+                );
+        }
+
+        const foundComments = isPlatformTypeGithub
+            ? reviewComments.filter((comment) =>
+                  implementedSuggestionsCommentIds.includes(
+                      Number(comment.fullDatabaseId),
+                  ),
+              )
+            : reviewComments.filter((comment) =>
+                  implementedSuggestionsCommentIds.includes(comment.id),
+              );
+
+        if (foundComments?.length > 0) {
+            const promises = foundComments.map(
+                async (foundComment: PullRequestReviewComment) => {
+                    let commentId =
+                        platformType === PlatformType.BITBUCKET
+                            ? foundComment.id
+                            : foundComment.threadId;
+
+                    return this.codeManagementService.markReviewCommentAsResolved(
+                        {
+                            organizationAndTeamData,
+                            repository,
+                            prNumber: pr.number,
+                            commentId: commentId,
+                        },
+                    );
+                },
+            );
+
+            // timeout mechanism for the Promise.allSettled operation to prevent potential hanging.
+            await Promise.race([
+                Promise.allSettled(promises),
+                new Promise((_, reject) =>
+                    setTimeout(
+                        () => reject(new Error('Operation timed out')),
+                        30000,
+                    ),
+                ),
+            ]);
+        }
+    }
+
+    private getImplementedSuggestionsCommentIds(
+        pr: PullRequestsEntity,
+    ): number[] {
+        const implementedSuggestionsCommentIds: number[] = [];
+
+        pr.files?.forEach((file) => {
+            if (file.suggestions.length > 0) {
+                file.suggestions
+                    ?.filter(
+                        (suggestion) =>
+                            suggestion.comment &&
+                            suggestion.implementationStatus !==
+                                ImplementationStatus.NOT_IMPLEMENTED &&
+                            suggestion.deliveryStatus === DeliveryStatus.SENT,
+                    )
+                    .forEach((filteredSuggestion) => {
+                        implementedSuggestionsCommentIds.push(
+                            filteredSuggestion.comment.id,
+                        );
+                    });
+            }
+        });
+
+        return implementedSuggestionsCommentIds;
+    }
+}
